Fix showing tracks of selected artist on 64-bit systems
The handler_id is of type gulong (while signal_id is guint); see
http://developer.gnome.org/gobject/unstable/gobject-Signals.html

Bug reported in http://sourceforge.net/mailarchive/message.php?msg_id=30490385

+++ plugins/sorttab_display/normal_sorttab_page.c
@@ -66,7 +66,7 @@ struct _NormalSortTabPagePrivate {
      * Handler id of the idle callback when the selection
      * changed callback is executed
      */
-    guint selection_changed_idle_id;
+    gulong selection_changed_idle_id;
 
     /* name of entry last selected */
     GList *last_selection;
